PILOT-84 | refactor for @extract_args
